// src/modules/permisos/enums/permisos-enum.ts
export enum PermisosEnum {
  ASIGNAR_RECLAMOS = 'ASIGNAR RECLAMOS',
  CREAR_USUARIOS = 'CREAR USUARIOS',
  ELIMINAR_USUARIOS = 'ELIMINAR USUARIOS',
  EDITAR_USUARIOS = 'EDITAR USUARIOS',
  MOVER_RECLAMO = 'MOVER RECLAMO A SUBÁREA / ÁREA / EMPLEADO',
  ASIGNAR_RECLAMO_A_EMPLEADO = 'ASIGNAR RECLAMO A EMPLEADO',
  CERRAR_RECLAMO = 'CERRAR RECLAMO',
  CREAR_PROYECTOS = 'CREAR PROYECTOS',
  ELIMINAR_PROYECTOS = 'ELIMINAR PROYECTOS',
  EDITAR_PROYECTOS = 'EDITAR PROYECTOS',
  AUTO_ASIGNAR_RECLAMO = 'AUTO-ASIGNAR RECLAMO',
  REGISTRAR_RECLAMO = 'REGISTRAR RECLAMO',
  VISUALIZAR_HISTORIAL = 'VISUALIZAR HISTORIAL',
  CREAR_FEEDBACK = 'CREAR FEEDBACK',
  VER_FEEDBACK = 'VER FEEDBACK',
  VISUALIZAR_ESTADO_RECLAMO = 'VISUALIZAR ESTADO DE RECLAMO',
  VER_RECLAMO = 'VER RECLAMO',
}